refactor: 당첨금 합계를 담는 변수의 자료형을 long으로 변경

+++ src/main/java/lotto/domain/WinningStatistics.java
@@ -23,9 +23,9 @@ public class WinningStatistics {
 
     public static double calculateEarningRate(Budget budget) {
         int amount = budget.getAmount();
-        int earnings = 0;
+        long earnings = 0;
         for (WinningInfo info : WinningInfo.values()) {
-            earnings += info.getPrizeMoney() * info.getWinningTicketCount();
+            earnings += (long) info.getPrizeMoney() * info.getWinningTicketCount();
         }
         return (double) earnings / amount * 100;
     }
